azure_functions_proj/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
djangorestframework

django~=3.0.0
django-environ

pyodbc
django-mssql-backend

django-storages[azure]
